feat: enhance task management with user ownership checks and access control; add custom exceptions for task access denial

+++ backend/app/services/tasks.py
@@ -1,4 +1,4 @@
-from asyncpg import InternalServerError
+from sqlalchemy.exc import IntegrityError
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy import select
 
@@ -10,83 +10,77 @@ from backend.app.dependencies import get_async_session
 from backend.app.models.tasks import Task as TaskModel
 
 from backend.app.schemas.tasks import TaskCreate, TaskUpdate
-from backend.app.services.statuses import get_status_by_id
-from backend.app.utils.custom_exceptions import TaskNotFoundException
+from backend.app.services.users import get_user_by_id
+from backend.app.utils.custom_exceptions import (
+    TaskNotFoundException,
+    TaskAccessDeniedException,
+    IntegrityErrorException,
+    InternalServerException,
+)
 
 # TODO: посмотреть, можно ли объявить свой тип, чтобы сократить Annotated[AsyncSession, Depends(get_async_session)]
 
 
 # TODO: сделать пагинацию задач
-async def get_all_tasks(session: Annotated[AsyncSession, Depends(get_async_session)]):
-    result = await session.scalars(select(TaskModel))
+async def get_all_tasks(
+    user_id: int, session: Annotated[AsyncSession, Depends(get_async_session)]
+):
+    result = await session.scalars(
+        select(TaskModel).where(TaskModel.owner_id == user_id)
+    )
     return result.all()
 
 
 async def get_task_by_id(
-    task_id: int, session: Annotated[AsyncSession, Depends(get_async_session)]
+    task_id: int,
+    user_id: int,
+    session: Annotated[AsyncSession, Depends(get_async_session)],
 ):
     task = (
         await session.scalars(select(TaskModel).where(TaskModel.id == task_id))
     ).first()
     if not task:
         raise TaskNotFoundException(f"Task with id {task_id} not found")
+    if task.owner_id != user_id:
+        raise TaskAccessDeniedException(
+            f"User with id {user_id} does not have permission to access task with id {task_id}"
+        )
     return task
 
 
 async def create_task(
-    task_data: TaskCreate, session: Annotated[AsyncSession, Depends(get_async_session)]
+    task_data: TaskCreate,
+    user_id: int,
+    session: Annotated[AsyncSession, Depends(get_async_session)],
 ):
-    task = TaskModel(**task_data.model_dump())
+    task_data = task_data.model_dump()
+    task_data["owner_id"] = user_id
+
+    executor_id = task_data.get("executor_id")
+    if executor_id is not None:
+        await get_user_by_id(executor_id, session)
+
+    task = TaskModel(**task_data)
     session.add(task)
     try:
         await session.commit()
         await session.refresh(task)
+    except IntegrityError:
+        await session.rollback()
+        raise IntegrityErrorException("Database integrity error")
     except Exception:
         await session.rollback()
-        raise InternalServerError("Unexpected error while creating task")
-    return task
-
-
-# TODO: нужна ли, если есть функция update_task?
-async def assign_executor_to_task(
-    task_id: int,
-    executor_id: int,
-    session: Annotated[AsyncSession, Depends(get_async_session)],
-):
-    task = await get_task_by_id(task_id, session)
-    task.executor_id = executor_id
-    try:
-        await session.commit()
-        await session.refresh(task)
-    except Exception:
-        await session.rollback()
-        raise InternalServerError("Unexpected error while assigning executor to task")
-    return task
-
-
-# TODO: нужна ли, если есть функция update_task?
-async def change_task_status(
-    task_id: int,
-    new_status_id: int,
-    session: Annotated[AsyncSession, Depends(get_async_session)],
-):
-    task = await get_task_by_id(task_id, session)
-    task.status_id = new_status_id
-    try:
-        await session.commit()
-        await session.refresh(task)
-    except Exception:
-        await session.rollback()
-        raise InternalServerError("Unexpected error while changing task status")
+        raise InternalServerException("Unexpected error while creating task")
     return task
 
 
 async def update_task(
     task_id: int,
     task_data: TaskUpdate,
+    user_id: int,
     session: Annotated[AsyncSession, Depends(get_async_session)],
 ):
-    task = await get_task_by_id(task_id, session)
+    task = await get_task_by_id(task_id, user_id, session)
     for key, value in task_data.model_dump(exclude_unset=True).items():
         setattr(task, key, value)
     try:
@@ -94,22 +88,30 @@ async def update_task(
         await session.refresh(task)
     except Exception:
         await session.rollback()
-        raise InternalServerError("Unexpected error while updating task")
+        raise InternalServerException("Unexpected error while updating task")
     return task
 
 
 async def delete_task(
-    task_id: int, session: Annotated[AsyncSession, Depends(get_async_session)]
+    task_id: int,
+    user_id: int,
+    session: Annotated[AsyncSession, Depends(get_async_session)],
 ):
-    task = await get_task_by_id(task_id, session)
+    task = await get_task_by_id(task_id, user_id, session)
     if not task:
         raise TaskNotFoundException(f"Task with id {task_id} not found")
+
+    owner_id = task.owner_id
+    if owner_id != user_id:
+        raise TaskAccessDeniedException(
+            f"User with id {user_id} does not have permission to delete task with id {task_id}"
+        )
     await session.delete(task)
     try:
         await session.commit()
     except Exception:
         await session.rollback()
-        raise InternalServerError("Unexpected error while deleting task")
+        raise InternalServerException("Unexpected error while deleting task")
     return {"detail": f"Task with id {task_id} deleted successfully"}
